requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
mysqlclient
python-dotenv
passlib[bcrypt]
google-auth
pyjwt
